requirements.txt
langchain==0.1.20
langchain-pinecone==0.1.0
langchain-community==0.0.38
langchain-aws==0.1.5
pinecone-client==3.2.2
boto3==1.34.108
pandas==2.2.2
numpy==1.26.4
neo4j==5.20.0